Restructure modules (#762)
* Move access module to top level

* Move graph module to top level

* Move test_utils to top level

* Make resolver module public

* Split out mod.rs

* fmt

* Move state.rs into crdt module

* Delete dgm.rs in group module

---------
