+++ p2panda-auth/src/lib.rs
@@ -1,4 +1,10 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+mod access;
+pub mod graph;
 pub mod group;
+#[cfg(any(test, feature = "test_utils"))]
+mod test_utils;
 pub mod traits;
+
+pub use access::{Access, AccessLevel};
