+++ p2panda-auth/src/test_utils/orderer.rs
@@ -10,12 +10,11 @@ use rand::rngs::StdRng;
 use thiserror::Error;
 
 use crate::group::{GroupAction, GroupControlMessage, GroupMember};
-use crate::traits::{GroupStore, Operation, Ordering};
-
-use super::{
+use crate::test_utils::{
     Conditions, MemberId, MessageId, PartialOrderer, PartialOrdererState, TestGroupState,
     TestGroupStore,
 };
+use crate::traits::{GroupStore, Operation, Ordering};
 
 #[derive(Debug, Error)]
 pub enum OrdererError {}
